Add handling for a partially-downloaded sticker pack
This can happen if the back button is pressed while
downloading/installing a pack, leaving a partial directory of sticker
images files that have not been registered with Firebase.
The .json file will not exist, so the pack will be listed as
installable.

The fix deletes the existing sticker pack directory if found.