Refactor ARActivity
This includes excising RecyclerView from the sticker pack
gallery, and just adding a bunch of ImageViews to a
HorizontalScrollView. This was because I want a rotation
animation when the screen rotates, and it looked like
getting a RecyclerView to animate elements in-place would
be a big hack, or at least a lot of boilerplate code.
I think loading up all the ImageViews initially is OK
ARCore is already loading a lot, so it's not a noticeable
lag, and it means things look a bit smoother when scrolling
quickly through a pack since everything is loaded already.