Prevent TextObject from constantly redrawing itself
Changing the TextObject's paint in the onDraw method seemed to be
invalidating the widget and requiring a re-draw the following frame,
causing constant rendering. Now the outline is drawn by a separate
TextView so that TextObject never has to change its paint and never
invalidates itself unnecessarily. This prevents a steady CPU usage
in EditorActivity of ~10% on the emulator, ~4% on a Pixel 3